Examples

Sunspots and Melanoma

The data can be obtained through the following helper function

AsymptoticPDC.get_sunspot_melanoma_data — Function

get_sunspot_melanoma_data()

This data is from Andrews and Herzberg. D. F. Andrews, A. M. Herzberg. (1985) Data: A Collection of Problems from Many Fields for the Student and Research Worker. Springer, New York.

Returns u:

u[:, 1]: Year 1936-1972

u[:, 2]: Annual male melanoma incidence (age-adjusted per 10e5) in Connecticut

u[:, 3]: Annual total melanoma incidence (age-adjusted per 10e5) in Connecticut

u[:, 4]: Annual sunspot relative number

Setup

A first look at the detrended data

u = get_sunspot_melanoma_data()
t = u[:, 1]
melanoma, _ = detrend(u[:, 3])
sunspots, _ = detrend(u[:, 4])
p1 = plot(t, melanoma, lab = "Detrended annual melanoma incidence Connecticut")
p2 = plot(t, sunspots, lab = "Detrended annual sunpot number")
plot(p1, p2, layout = (2, 1))

Estimation

Estimate the model parameters using mcar

y = [melanoma sunspots]
model, _, _ = mcar(y)

(MCAR_Model{Float64, Array{Float64, 3}, Matrix{Float64}}(3, 2, 37, [-0.15136402462500836 -0.0005913956663627665; -13.007134576783498 0.765652335488139;;; -0.0762157476017296 0.004227477213695851; 16.6699493203485 -0.04394536617446674;;; 0.09778093472058648 0.002891647014980567; -28.62446855694505 -0.2832230418733722;;; -0.10327559460573496 -0.0028936913155264863; -3.578229944054352 -0.13644990698633863], [0.04270223341659899 -1.7919220155852729; -1.7919220155852633 558.6072604240353], [0.07709815078236182 -24.004267425320066; -0.06266761070267378 74.64659715915631; … ; -0.17371639273352027 2.375269026946955; -0.0403663923409128 -2.745645999156643]), 142.24657012533189, [150.9486885199741, 148.14305646442727, 142.24657012533189, 144.0125514977107, 0.0, 0.0, 0.0, 0.0, 0.0])

Connectivity

Perform a Granger causality test to get a connectivity matrix

connectivity, pvalues = granger_causality_test(model, y);

                  GRANGER CAUSALITY TEST
======================================================================
Connectivity matrix:

Granger causality test p-values:

======================================================================
Connectivity matrix:
[NaN 1.0; 0.0 NaN]
Granger causality test p-values:
[NaN 5.388086010693627e-6; 0.3426760795791506 NaN]

Original PDC analysis

original_pdc = pdc(model, y; metric = "euc")
pdcplot(original_pdc, cnames = ["Melanoma", "Sunspots"])

Generalized PDC analysis

generalized_pdc = pdc(model, y; metric = "diag")
pdcplot(generalized_pdc, cnames = ["Melanoma", "Sunspots"])