Multichannel Autoregression Model

Definition: A multichannel or multivariate autoregressive model (MCAR) of order $p$ is defined as

\[X(t) = \sum_{i=1}^{p} A(i)X(t-i) + E(t)\]

where $X(t) = [X_1(t), X_2(t), ...,X_k(t)]$ is a k-channel set of signals and $E(t)$ a vector of k withe noises at time $t$. $A$ is a $k \times k \times p$ matrix holding the model parameters.

AsymptoticPDC.mcar — Function

mcar(u; maxorder::Union{Nothing,Int}=nothing, criterion::Union{Nothing,String}="AIC", method::String="NS", verbose::Bool=true)

Compute a multichannel AR or vector AR model of the input matrix u containing the signals/channels xi, u = [x1 x2 ... xn]

Args

u: input Matrix containing signals 1 to n u = [x1 x2 ... xn]

Keywords

maxorder::Union{Nothing, Int} = nothing: The maximal order of the AR model, defaults to nothing where the order is chosen based on a simple heuristic (maxorder = 3√samples/nChannels; Nuttall 1976)

criterion = "AIC": The information criterion used to choose the model order. Use one of the following:

"AIC": Akaike's Informaion Criterion

"HQ": Hannan Quinn

"BIC": Bayesian Information Criterion, Schwarz 1978

"FPE": Final prediction error, Akaike, 1970

nothing: maxorder becomes the fixed order

method = "LS": Method used for etsimation. Use one of:

"LS" least squares based on \

"NS" Nuttall-Strand Method (multi-channel generalization of the single-channel Burg lattice algorithm)

"VM" Vieira-Morf Method (multi-channel generalization of the single-channel geometric lattice algorithm)

Return

Returns a tuple (model, besticvalue, ic_values)

Result model is of type MCAR_Model, with following fields:

order: is the (chosen) model order

nChannels: number of channels

samples: number of samples per channel

A: contains the AR coefficients [n x n x order]

pf: is the covariance matrix [order x order]

ef: the residuals

AsymptoticPDC.MCAR_Model — Type

MCAR_Model with following fields:

order: is the (chosen) model order

nChannels: number of channels

samples: number of samples per channel

A: contains the AR coefficients [n x n x order]

pf: is the covariance matrix [order x order]

ef: the residuals