Examples

Sunspots and Melanoma

The data can be obtained through the following helper function

AsymptoticPDC.get_sunspot_melanoma_data — Function

get_sunspot_melanoma_data()

This data is from Andrews and Herzberg. D. F. Andrews, A. M. Herzberg. (1985) Data: A Collection of Problems from Many Fields for the Student and Research Worker. Springer, New York.

Returns u:

u[:, 1]: Year 1936-1972

u[:, 2]: Annual male melanoma incidence (age-adjusted per 10e5) in Connecticut

u[:, 3]: Annual total melanoma incidence (age-adjusted per 10e5) in Connecticut

u[:, 4]: Annual sunspot relative number

Setup

A first look at the detrended data

u = get_sunspot_melanoma_data()
t = u[:, 1]
melanoma, _ = detrend(u[:, 3])
sunspots, _ = detrend(u[:, 4])
p1 = plot(t, melanoma, lab = "Detrended annual melanoma incidence Connecticut")
p2 = plot(t, sunspots, lab = "Detrended annual sunpot number")
plot(p1, p2, layout = (2, 1))

Estimation

Estimate the model parameters using mcar

y = [melanoma sunspots]
model, _, _ = mcar(y)

(MCAR_Model{Float64, Array{Float64, 3}, Matrix{Float64}}(3, 2, 37, [-0.15136402462500836 -0.0005913956663627665; -13.007134576783498 0.765652335488139;;; -0.0762157476017296 0.004227477213695851; 16.6699493203485 -0.04394536617446674;;; 0.09778093472058648 0.002891647014980567; -28.62446855694505 -0.2832230418733722;;; -0.10327559460573496 -0.0028936913155264863; -3.578229944054352 -0.13644990698633863], [0.04270223341659899 -1.7919220155852729; -1.7919220155852633 558.6072604240353], [0.07709815078236182 -24.004267425320066; -0.06266761070267378 74.64659715915631; … ; -0.17371639273352027 2.375269026946955; -0.0403663923409128 -2.745645999156643]), 142.24657012533189, [150.9486885199741, 148.14305646442727, 142.24657012533189, 144.0125514977107, 0.0, 0.0, 0.0, 0.0, 0.0])

Connectivity

Perform a Granger causality test to get a connectivity matrix

connectivity, pvalues = granger_causality_test(model, y);
println("Connectivity matrix:")
display(connectivity)
println("Granger causality test p-values:")
display(pvalues)

                  GRANGER CAUSALITY TEST
======================================================================
Connectivity matrix:

Granger causality test p-values:

======================================================================
Connectivity matrix:
Granger causality test p-values:

Original PDC analysis

original_pdc = pdc(model, y; metric = "euc")

AsymptoticPDC.PartialDirectedCoherence{Float64, StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}, Array{Float64, 3}, Array{ComplexF64, 3}}(0.0:0.003937007874015748:0.5, [0.0018632204597483122 2.7108061747300376e-5; 0.9981367795402518 0.9999728919382527;;; 0.0018637197015387678 2.7236241396525112e-5; 0.9981362802984612 0.9999727637586034;;; 0.0018652247620307432 2.7624329923465e-5; 0.9981347752379693 0.9999723756700766;;; … ;;; 0.00041967671826979966 4.803208594327655e-7; 0.9995803232817303 0.9999995196791407;;; 0.0004215925903126524 4.011875683595906e-7; 0.9995784074096873 0.9999995988124316;;; 0.0004227475068545509 3.5366394245081784e-7; 0.9995772524931456 0.9999996463360575], [0.02052668847020703 + 0.0im 0.31845562034141683 + 0.0im; 0.3184556203414268 + 0.0im 1086.3050438193281 + 0.0im;;; 0.02055177115035817 - 2.0328790734103208e-19im 0.3260579960411879 - 0.07275027510860765im; 0.3260579960411971 + 0.07275027510860776im 1089.1987942362127 - 2.1316282072803006e-14im;;; 0.020628110222852464 - 4.607859233063394e-19im 0.34906660825101554 - 0.14778252947650888im; 0.3490666082510261 + 0.1477825294765091im 1097.958695143956 - 5.3290705182007514e-14im;;; … ;;; 0.03417363610526172 - 3.469446951953614e-18im 0.018259435386795753 - 0.23883493149889276im; 0.01825943538680054 + 0.2388349314988928im 168.617248631561 - 1.865174681370263e-14im;;; 0.033828712764984135 - 2.3310346708438345e-18im 0.010746736075554875 - 0.1580274011075964im; 0.010746736075559815 + 0.15802740110759636im 167.16082819068976 - 1.1546319456101628e-14im;;; 0.03362328014790984 - 1.0028870095490916e-18im 0.006285172311006448 - 0.07865561065170323im; 0.006285172311011028 + 0.07865561065170323im 166.29198090237998 - 5.773159728050814e-15im])